+++ packages/ui/src/components/primitives/ChainIcon.tsx
@@ -19,7 +19,7 @@ const ChainIcon: FC<Props> = ({
   height = 14,
   width = 14,
   square = true,
-  borderRadius
+  borderRadius = 4
 }) => {
   const providerOptions = useContext(ProviderOptionsContext)
   const client = useRelayClient()
@@ -45,7 +45,8 @@ const ChainIcon: FC<Props> = ({
       }}
       style={{
         height: height,
-        width: width
+        width: width,
+        borderRadius: borderRadius
       }}
     >
       {iconUrl ? (
@@ -53,7 +54,7 @@ const ChainIcon: FC<Props> = ({
           src={iconUrl}
           alt={`Chain #${chainId}`}
           style={{
-            borderRadius: borderRadius ?? (square ? 4 : 0),
+            borderRadius: square ? borderRadius : 0,
             width: css && css.width ? Number(css.width) : '100%',
             height: css && css.height ? Number(css.height) : '100%'
           }}