+++ packages/ui/src/components/widgets/SwapWidget/index.tsx
@@ -66,6 +66,7 @@ type BaseSwapWidgetProps = {
   popularChainIds?: number[]
   disablePasteWalletAddressOption?: boolean
   sponsoredTokens?: string[]
+  onOpenSlippageConfig?: () => void
   onFromTokenChange?: (token?: Token) => void
   onToTokenChange?: (token?: Token) => void
   onConnectWallet?: () => void
@@ -103,6 +104,7 @@ const SwapWidget: FC<SwapWidgetProps> = ({
   defaultAmount,
   defaultTradeType,
   slippageTolerance,
+  onOpenSlippageConfig,
   lockToToken = false,
   lockFromToken = false,
   lockChainId,
@@ -251,6 +253,7 @@ const SwapWidget: FC<SwapWidgetProps> = ({
         gasTopUpEnabled,
         setGasTopUpEnabled,
         gasTopUpRequired,
+        gasTopUpBalance,
         gasTopUpAmount,
         gasTopUpAmountUsd,
         linkedWallet,
@@ -1447,6 +1450,23 @@ const SwapWidget: FC<SwapWidgetProps> = ({
                           popularChainIds={popularChainIds}
                         />
                       </Flex>
+                      <GasTopUpSection
+                        toChain={toChain}
+                        gasTopUpEnabled={gasTopUpEnabled}
+                        onGasTopUpEnabled={(enabled) => {
+                          setGasTopUpEnabled(enabled)
+                          onAnalyticEvent?.(EventNames.GAS_TOP_UP_TOGGLE, {
+                            enabled,
+                            amount: gasTopUpAmount,
+                            amount_usd: gasTopUpAmountUsd,
+                            quote
+                          })
+                        }}
+                        gasTopUpRequired={gasTopUpRequired}
+                        gasTopUpAmount={gasTopUpAmount}
+                        gasTopUpAmountUsd={gasTopUpAmountUsd}
+                        gasTopUpBalance={gasTopUpBalance}
+                      />
                       <Flex
                         align="center"
                         justify="between"
@@ -1567,22 +1587,6 @@ const SwapWidget: FC<SwapWidgetProps> = ({
                         />
                       </Box>
                     ) : null}
-                    <GasTopUpSection
-                      toChain={toChain}
-                      gasTopUpEnabled={gasTopUpEnabled}
-                      onGasTopUpEnabled={(enabled) => {
-                        setGasTopUpEnabled(enabled)
-                        onAnalyticEvent?.(EventNames.GAS_TOP_UP_TOGGLE, {
-                          enabled,
-                          amount: gasTopUpAmount,
-                          amount_usd: gasTopUpAmountUsd,
-                          quote
-                        })
-                      }}
-                      gasTopUpRequired={gasTopUpRequired}
-                      gasTopUpAmount={gasTopUpAmount}
-                      gasTopUpAmountUsd={gasTopUpAmountUsd}
-                    />
                     <FeeBreakdown
                       feeBreakdown={feeBreakdown}
                       isFetchingQuote={isFetchingQuote}
@@ -1605,6 +1609,9 @@ const SwapWidget: FC<SwapWidgetProps> = ({
                       isSingleChainLocked={isSingleChainLocked}
                       fromChainWalletVMSupported={fromChainWalletVMSupported}
                       error={error}
+                      onOpenSlippageConfig={() => {
+                        onOpenSlippageConfig?.()
+                      }}
                     />
                     <WidgetErrorWell
                       hasInsufficientBalance={hasInsufficientBalance}